+++ alembic/env.py
@@ -19,6 +19,8 @@ import app.domain.store.models  # noqa: F401
 import app.domain.product.models # noqa: F401
 import app.domain.order.models   # noqa: F401
 import app.domain.delivery.models # noqa: F401
+import app.domain.notification.models # noqa: F401
+
 
 
 config = context.config