+++ app/main.py
@@ -20,6 +20,8 @@ from app.domain.user.routes import router as user_router
 from app.domain.freight.routes import router as freight_router
 from app.domain.order.routes import router as order_router
 from app.domain.delivery.routes import router as delivery_router
+from app.domain.notification.routes import router as notification_router
+
 
 
 
@@ -168,4 +170,7 @@ app.include_router(order_router, prefix=settings.API_V1_STR)
 app.include_router(order_router, include_in_schema=False)
 app.include_router(delivery_router, prefix=settings.API_V1_STR)
 app.include_router(delivery_router, include_in_schema=False)
+app.include_router(notification_router, prefix=settings.API_V1_STR)
+app.include_router(notification_router, include_in_schema=False)
+